// src/api/generated/models/PipelineStageResponse.ts
export interface PipelineStageResponse {
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    stageId?: string;
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    stageName?: string;
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    stageCode?: string;
    /**
     * 
     * @type {number}
     * @memberof PipelineStageResponse
     */
    stageOrder?: number;
    /**
     * 
     * @type {number}
     * @memberof PipelineStageResponse
     */
    probability?: number;
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    stageColor?: string;
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    isWon?: string;
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    isLost?: string;
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof PipelineStageResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof PipelineStageResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PipelineStageResponse
     */
    updatedDate?: Date;
}